+++ mods/Veil/common/src/main/java/foundry/veil/api/client/render/shader/block/DynamicShaderBlock.java
@@ -11,12 +11,12 @@ public interface DynamicShaderBlock<T> extends ShaderBlock<T> {
     /**
      * @return The size of this block in bytes
      */
-    long getSize();
+    int getSize();
 
     /**
      * Resizes this shader block to match the new size.
      *
      * @param newSize The size in bytes
      */
-    void setSize(long newSize);
+    void setSize(int newSize);
 }
